Make git depend less on human-readable output
The Debian version of Git respects the locale, which ends up confusing Git.pm as it looks for "# On branch ..." , where my German localized version of Git outputs "Auf Zweig ..." . The two changes change the following:

1) Use the short status of git

This only outputs a list of pending changes. If that list is not empty, we want to bail.

2) Use `git rev-parse --abbrev-ref HEAD` to find the current branch.

https://stackoverflow.com/questions/6245570/how-to-get-current-branch-name-in-git suggests that is
is the best way of finding the currently active branch.

Thanks for maintaining Module::Release and Module::Release::Git!