Update clean working dir output for recent gits
It turns out that the "# On branch" output got changed to "On branch" in
git-1.5.0 and that the "working directory clean" got changed to "working
tree clean" in git-2.9.1.  This change updates the code to handle the
new output when a recent git is used when using `Module::Release` to
release a module.  The original behaviour is retained and the tests have
been extended to check the newer behaviour.