Add method to return list of recent contributors
This is the concrete implementation of the stub method of the same name
in `Module::Release`.  It searches for the most recent git tag, finds
the commit corresponding to this tag and then returns author information
for all commits between the HEAD commit and the most recent tag (which
is assumed to also be the tag of the most recent release).